+++ src/main/java/tdd/constants/BoardPositions.java
@@ -1,0 +1,13 @@
+package tdd.constants;
+
+public class BoardPositions {
+    public static final int POSITION_TOP_LEFT = 0;
+    public static final int POSITION_TOP = 1;
+    public static final int POSITION_TOP_RIGHT = 2;
+    public static final int POSITION_LEFT = 3;
+    public static final int POSITION_CENTER = 4;
+    public static final int POSITION_RIGHT = 5;
+    public static final int POSITION_BOTTOM_LEFT = 6;
+    public static final int POSITION_BOTTOM = 7;
+    public static final int POSITION_BOTTOM_RIGHT = 8;
+}